requirements.txt
google-generativeai
python-dotenv~=1.1.0
protobuf~=5.29.4
beautifulsoup4
requests
python-telegram-bot==20.8